django_app/requirements/base.txt
Django==5.0.4
djangorestframework==3.15.2

# Swagger
drf-spectacular==0.27.2

# Filtering
django-filter==24.3

# postgresql
psycopg[binary]==3.1.18

# read .env
python-dotenv==1.0.1

# split settings.py
split-settings==1.0.0

# Translation
django-rosetta==0.10.0